Спринт 19. Заменил MainActivity на MainScreenFragment.
