+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         tools:targetApi="31">
+
         <activity
             android:name=".search.presentation.SearchActivity"
             android:exported="false"
@@ -23,6 +24,7 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
+
         <activity
             android:name=".audioplayer.presentation.AudioPlayerActivity"
             android:exported="false"
@@ -31,6 +33,7 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
+
         <activity
             android:name=".medialibrary.presentation.MediaLibrary"
             android:exported="false"
@@ -39,6 +42,7 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
+
         <activity
             android:name=".settings.presentation.SettingsActivity"
             android:exported="false"
@@ -47,20 +51,16 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
-        <activity
-            android:name=".mainscreen.presentation.MainActivity"
-            android:exported="true"
-            android:theme="@style/Theme.PlaylistMaker.MainMenu">
+
+        <activity android:name=".RootActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
         </activity>
+
     </application>
 
 </manifest>
