+++ app/src/main/java/com/practicum/playlistmaker/RootActivity.kt
@@ -1,0 +1,26 @@
+package com.practicum.playlistmaker
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.commit
+import com.practicum.playlistmaker.databinding.ActivityRootBinding
+import com.practicum.playlistmaker.mainscreen.presentation.MainScreenFragment
+
+class RootActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityRootBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityRootBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        if (savedInstanceState == null) {
+            supportFragmentManager.commit {
+                this.add(R.id.rootFragmentContainerView, MainScreenFragment())
+            }
+        }
+    }
+
+}
