+++ app/src/main/res/layout/activity_root.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/rootFragmentContainerView"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent" />
