+++ app/src/main/res/layout/fragment_main.xml
@@ -3,7 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
+    android:orientation="vertical"
+    android:theme="@style/Theme.PlaylistMaker.MainMenu"
+    >
 
     <TextView
         android:layout_width="wrap_content"
